+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -128,6 +128,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added an initial version of SSGI.
 - Added back-compatibility with builtin stereo matrices.
 - Added CustomPassUtils API to simplify Blur, Copy and DrawRenderers custom passes.
+- Added Histogram guided automatic exposure.
+- Added few exposure debug modes.
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)